requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
delegator.py==0.1.1
docker==5.0.0
english==2020.7.0
idna==2.10
monotonic==1.6
neotime==1.7.4
packaging==20.9
pansi==2020.7.3
pexpect==4.8.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py2neo==2021.1.3
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
python-magic==0.4.23
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
wcwidth==0.2.5
websocket-client==1.0.1
